Create Alzheimer detection.py
the advancement of Alzheimer's illness utilizing Convolution Neural Systems (CNN) and EfficientNetB3 architecture, which was applied to pre-processed MRI datasets. The purpose of the project is to use efficient high-performance computing (HPC) to improve the performance of the model, which makes the diagnostic process more efficient and reliable the result of this project is Training time without HPC optimizations: 976.88 seconds. Training time with HPC optimizations: 635.31 seconds.